Streams rewrite
This is a major change.  To reading, the poller and the bindings layer, there isn’t a great way to do this piecemeal. The goal here is to make bindings a plugin layer, our c++ bindings isolated and tested, and make serialport a valid streams object.

- Move all bindings into platform specific files
- Provide mock bindings as a tested first class bindings layer
- Separate out the binary bindings from the SerialPort JS so it can be required without bindings
- Make Bindings manage .isOpen and FD as file descriptors are an implementation detail
- Make Bindings an instantiated object so they can isolate their own state, have setup code, etc.
- Since all integration tests now require an Arduino ditch the "integration-light" tests and test with both native and mock bindings
- update docs to match new stream methods
- ask mocha to fail fast
- implement `read()` spec for the bindings layer
- mixin stream methods with `pushBindingWrap`


TODO

 - finish osx binding
 - finish windows binding
 - finish linux binding
 - ensure bindings are documented
 - performance test
 - 🎉

The c++ bindings use .open to create some global state around write queues and read pollers. Now that the stream will only have a single call to write out at any time this is not necessary. In any case, hiding this state in a “global scope” is bad practice.

The bindings `read()` function is very very inefficient as it can't make use of a "buffer pool" and allocates a full buffer every read regardless of how much it can actually read. This will probably have to change to be a lot more like `fs.read`